+++ src/main/java/com/aiot/aiotbackstage/mapper/SysInsectRecMapper.java
@@ -18,5 +18,5 @@ public interface SysInsectRecMapper extends BaseMapper<SysInsectRecEntity> {
     List<SysInsectRecStatisEntity> findAllPestNumHourly(Map<String, Object> params);
 
 
-    List<SysInsectRecEntity> insectRecGisInfo(@Param("site_id") Long site_id);
+    List<SysInsectRecEntity> insectRecGisInfo(@Param("site_id") Integer site_id);
 }
