+++ src/main/java/com/aiot/aiotbackstage/service/IGisService.java
@@ -19,7 +19,7 @@ public interface IGisService {
 
     void disasterSituation(DisasterSituationGisParam param);
 
-    SensorInfoVo sensorInfo(Long stationId);
+    SensorInfoVo sensorInfo(Integer stationId);
 
     PageResult<SysSeedlingGrowthEntity> getSeedlingGrowth(PageParam pageParam);
